Rename Linux UART driver to POSIX (nasa/fprime#5698)
