+++ Drv/CMakeLists.txt
@@ -11,6 +11,7 @@ add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/ByteStreamDriverModel/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxGpioDriver/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxI2cDriver/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxSpiDriver/")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/PosixUartDriver/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/LinuxUartDriver/")
 
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Ip/")
