+++ Drv/LinuxUartDriver/CMakeLists.txt
@@ -1,19 +1,14 @@
-####
-# F prime CMakeLists.txt:
-#
-# SOURCE_FILES: combined list of source and autocoding files
-# MOD_DEPS: (optional) module dependencies
-#
-####
-restrict_platforms(Linux Darwin)
+restrict_platforms(Posix)
+
+if (NOT FPRIME_IS_SUB_BUILD)
+    message(DEPRECATION "Drv_LinuxUartDriver is deprecated; use Drv_PosixUartDriver")
+endif()
 
 register_fprime_module(
-  AUTOCODER_INPUTS
-    "${CMAKE_CURRENT_LIST_DIR}/LinuxUartDriver.fpp"
   SOURCES
     "${CMAKE_CURRENT_LIST_DIR}/LinuxUartDriver.cpp"
   HEADERS
     "${CMAKE_CURRENT_LIST_DIR}/LinuxUartDriver.hpp"
   DEPENDS
-    Os
+    Drv_PosixUartDriver
 )
